+++ CHANGELOG.md
@@ -26,6 +26,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - fixed #106
 - fix: use pcall to clear autocmd silently projekt0n/github-nvim-theme#178 ( @xiyaowong )
 - fixed #108
+- fixed: pcall bad argument inside autocmds module
 
 ## [v0.0.4]- 21 Jan 2022